Move Remus.TypeParsing -> Remus.Parsers.TypeParsers
