+++ src/Remus/Parsing/TypeParsers/ByteParser.cs
@@ -1,4 +1,4 @@
-﻿namespace Remus.TypeParsing
+﻿namespace Remus.Parsing.TypeParsers
 {
     internal sealed class ByteParser : ITypeParser<byte>
     {
