+++ src/Remus/Parsing/TypeParsers/ITypeParser.cs
@@ -1,4 +1,4 @@
-﻿namespace Remus.TypeParsing
+﻿namespace Remus.Parsing.TypeParsers
 {
     /// <summary>
     ///     Describes a type parser.
