+++ src/Remus/Parsing/TypeParsers/Int16Parser.cs
@@ -1,4 +1,4 @@
-﻿namespace Remus.TypeParsing
+﻿namespace Remus.Parsing.TypeParsers
 {
     internal sealed class Int16Parser : ITypeParser<short>
     {
