+++ src/Remus/Parsing/TypeParsers/Int64Parser.cs
@@ -1,4 +1,4 @@
-﻿namespace Remus.TypeParsing
+﻿namespace Remus.Parsing.TypeParsers
 {
     internal sealed class Int64Parser : ITypeParser<long>
     {
