+++ src/Remus/Parsing/TypeParsers/UInt32Parser.cs
@@ -1,4 +1,4 @@
-﻿namespace Remus.TypeParsing
+﻿namespace Remus.Parsing.TypeParsers
 {
     internal sealed class UInt32Parser : ITypeParser<uint>
     {
